refactor: §3.5 split validations.ts (819 LOC) into domain modules
Split the monolithic validations.ts into 9 domain-specific modules:
- validations/primitives.ts  — safeText, safeName, normalizeText, isoDate, timeHHMM
- validations/booking.ts     — booking, scheduling, waiting list, checkin
- validations/payments.ts    — payments, Stripe/CMI webhooks, billing
- validations/admin.ts       — onboarding, branding, custom fields, consent
- validations/chat.ts        — chat, AI schemas (prescription, drug check, etc.)
- validations/clinical.ts    — lab, radiology, upload, pet profiles
- validations/restaurant.ts  — menu, tables, orders
- validations/v1.ts          — V1 public API schemas
- validations/notifications.ts — notification dispatch/trigger
- validations/helpers.ts     — safeParse utility

The original validations.ts is now a barrel re-export (11 LOC) so all
54 existing import sites continue to work unchanged.
